Pulled out path flag for migration commands so it could be applied separately
