+++ .spelling
@@ -492,6 +492,7 @@ MatchReasonCodeFull
 MatchReasonCodeNone
 wkEma
 _No_
+subcommands
 # Put all custom terms BEFORE this comment, lest 'pre-commit' and 'make spellcheck' yield different errors.
  - /usr/local
  - docs/data/tspp-data-creation.md
