+++ pkg/cli/migration_path_test.go
@@ -1,0 +1,6 @@
+package cli
+
+func (suite *cliTestSuite) TestConfigMigrationPath() {
+	suite.Setup(InitMigrationPathFlags, []string{})
+	suite.NoError(CheckMigrationPath(suite.viper))
+}
